Refactor: UserContent 컴포넌트 모듈화 개선
